refactor: rename base search request class
