+++ clients/java/src/main/java/io/camunda/client/api/search/sort/ProcessDefinitionSort.java
@@ -15,7 +15,7 @@
  */
 package io.camunda.client.api.search.sort;
 
-import io.camunda.client.api.search.query.TypedSearchQueryRequest.SearchRequestSort;
+import io.camunda.client.api.search.query.TypedSearchRequest.SearchRequestSort;
 
 public interface ProcessDefinitionSort extends SearchRequestSort<ProcessDefinitionSort> {
