+++ clients/java/src/main/java/io/camunda/client/api/search/sort/VariableSort.java
@@ -15,7 +15,7 @@
  */
 package io.camunda.client.api.search.sort;
 
-import io.camunda.client.api.search.query.TypedSearchQueryRequest.SearchRequestSort;
+import io.camunda.client.api.search.query.TypedSearchRequest.SearchRequestSort;
 
 public interface VariableSort extends SearchRequestSort<VariableSort> {
   VariableSort variableKey();
